requirements.txt
flask
openai==0.28
pandas==1.5.3
numpy==1.23.2
openpyxl
pandasai
azure-storage-blob
opentelemetry-sdk==1.32.1
azure-ai-ml
azure-identity
transformers
datasets
tokenizers
torch
tqdm
accelerate
python-dotenv
ipykernel
peft